refactor(client): share the tile grid, KPI primitives, and work queues
The home screen needs the same tile canvas the report dashboard already has,
the same KPI cards the shipment page already has, and the same work queues the
command center already has. Duplicating any of them would guarantee they drift,
so each moves to a shared location and both consumers read the one copy.

- components/tile-grid/ takes the dnd-kit sortable grid out of
  dashboard-view.tsx: packTiles is now generic over {id, w, h}, SortableTile
  owns the grid spans and drag transform, and SizeStepper is the shared resize
  control. SortableTile hands the drag props to its children through a render
  prop — calling useSortable twice for one id would leave dnd-kit holding the
  wrong node and silently break dragging.

- components/kpi/ takes the KPI primitives out of routes/shipment. The Tone
  type moves with them: it belongs to the KPI vocabulary, not to any one page's
  data shape. kpi-rail stays behind, since its card set is specific to the
  shipment page.

- components/work-modules/ takes the unassigned and exceptions queues out of
  the command center's right stack, split into chrome-free list bodies that
  take an onSelect callback. The command center still supplies its own module
  card and its own URL behaviour; the home screen will supply different ones.
  lane-heatmap and customer-mix move as-is, and the shipment analytics payload
  shape moves to lib/ where both routes can reach it.

The report dashboard and the shipment command center render exactly as before:
the test suite is unchanged at 350 passing, plus 10 new cases covering the
extracted grid packing.
