+++ client/apps/web/src/components/kpi/__tests__/kpi-ring.test.tsx
