+++ client/apps/web/src/components/kpi/kpi-goal-bar.tsx
