+++ client/apps/web/src/components/kpi/kpi-ring.tsx
