+++ client/apps/web/src/components/kpi/segmented-bar.tsx
