+++ client/apps/web/src/routes/admin/api-keys/_components/analytics/cards/revoked-keys-card.tsx
@@ -1,4 +1,4 @@
-import { KPICard } from "@/routes/shipment/_components/analytics/kpi-card";
+import { KPICard } from "@/components/kpi/kpi-simple-card";
 import { ShieldOff } from "lucide-react";
 import type { ApiKeyAnalyticsData } from "../mock-data";
