+++ client/apps/web/src/routes/shipment/_components/command-center/cells/margin-cell.tsx
@@ -1,7 +1,7 @@
 import { getMarginTone, parseDecimal, resolveTargetMarginPct } from "@/lib/profitability";
 import { formatCurrency, formatPercent } from "@trenova/shared/lib/utils";
 import type { Shipment } from "@trenova/shared/types/shipment";
-import { toneVar } from "../../analytics/kpi/tone";
+import { toneVar } from "@/components/kpi/tone";
 import { ProfitabilityBreakdownPopover } from "../../profitability/profitability-breakdown-popover";
 
 export function MarginCell({ shipment }: { shipment: Shipment }) {
